new test database applied

+++ src/main/java/com/xiao/scooerproject/repository/PlayerDAO.java
@@ -6,9 +6,10 @@ import com.xiao.soccerproject.model.Team;
 import java.util.List;
 
 public interface PlayerDAO {
+//    boolean update(Player players);
     boolean save(Player players);
-    int updatePlayerAge(int id, int age);
-    int deleteById(int id);
+    int updatePlayerAge(long id, int age);
+    int deleteById(long id);
     List<Player> getPlayers();
-    Player getPlayerById(int id);
+    Player getPlayerById(long id);
 }
