+++ src/main/java/com/xiao/soccerproject/model/Team.java
@@ -12,7 +12,7 @@ public class Team {
     private String teamName;
     @Id
     @Column(name="id")
-    private int id;
+    private long id;
 
     @Column(name = "home_win")
     private int homeWin;
@@ -26,40 +26,6 @@ public class Team {
     @Column(name = "away_loss")
     private int awayLoss;
 
-//    @Column(name = "matches_won")
-//    private int matchesWon;
-//
-//    @Column(name = "matches_lost")
-//    private int matchesLost;
-//
-//    @Column(name = "matches_drawn")
-//    private int matchesDrawn;
-//
-//    @Column(name = "total_matches")
-//    private int totalMatches;
-//
-//    @Column(name = "points")
-//    private int points;
-//
-//    @Column(name = "finish_position")
-//    private int finishPosition;
-//
-//    @Column(name = "home_goals")
-//    private int homeGoals;
-//
-//    @Column(name = "away_goals")
-//    private int awayGoals;
-//
-//    @Column(name = "goals_scored")
-//    private int goalsScored;
-//
-//    @Column(name = "goals_conceded")
-//    private int goalsConceded;
-//
-//    @Column(name = "goal_difference")
-//    private int goalDifference;
-
-
     public String getTeamName() {
         return teamName;
     }
@@ -68,11 +34,11 @@ public class Team {
         this.teamName = teamName;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
